+++ test/models/artefact_action_test.rb
@@ -39,86 +39,96 @@ class ArtefactActionTest < ActiveSupport::TestCase
     }
   end
 
-  test "a new artefact should have a create action" do
-    a = Artefact.create!(base_fields)
-    a.reload
+  def setup
+    @artefact = Artefact.create!(base_fields)
+  end
 
-    assert_equal 1, a.actions.size
-    action = a.actions.first
+  test "a new artefact should have a create action" do
+    @artefact.reload
+
+    assert_equal 1, @artefact.actions.size
+    action = @artefact.actions.first
     assert_equal "create", action[:action_type]
     assert_equal merge_attributes(DEFAULTS, base_fields), action.snapshot
     assert action.created_at, "Action has no creation timestamp"
   end
 
   test "an updated artefact should have two actions" do
-    a = Artefact.create!(base_fields)
-    a.description = "An artefact of shining wonderment."
-    a.save!
-    a.reload
+    @artefact.description = "An artefact of shining wonderment."
+    @artefact.save!
+    @artefact.reload
 
-    assert_equal 2, a.actions.size
-    assert_equal ["create", "update"], a.actions.map(&:action_type)
+    assert_equal 2, @artefact.actions.size
+    assert_equal ["create", "update"], @artefact.actions.map(&:action_type)
     create_snapshot = merge_attributes(DEFAULTS, base_fields)
-    update_snapshot = create_snapshot.merge("description" => a.description)
-    assert_equal create_snapshot, a.actions[0].snapshot
-    assert_equal update_snapshot, a.actions[1].snapshot
-    a.actions.each do |action|
+    update_snapshot = create_snapshot.merge("description" => @artefact.description)
+    assert_equal create_snapshot, @artefact.actions[0].snapshot
+    assert_equal update_snapshot, @artefact.actions[1].snapshot
+    @artefact.actions.each do |action|
       assert action.created_at, "Action has no creation timestamp"
     end
   end
 
   test "saving with no tracked changes will not create a new snapshot" do
-    a = Artefact.create!(base_fields)
-    a.updated_at = Time.zone.now + 5.minutes
-    a.save!
-    assert_equal 1, a.actions.size
+    @artefact.updated_at = Time.zone.now + 5.minutes
+    @artefact.save!
+    assert_equal 1, @artefact.actions.size
   end
 
   test "updating attributes as a user should record a user action" do
-    a = Artefact.create!(base_fields)
     user = FactoryGirl.create :user
     updates = {description: "Shiny shiny description"}
-    a.update_attributes_as user, updates
-    a.reload
+    @artefact.update_attributes_as user, updates
+    @artefact.reload
 
-    assert_equal "Shiny shiny description", a.description
-    assert_equal 2, a.actions.size
-    assert_equal ["create", "update"], a.actions.map(&:action_type)
-    assert_equal user, a.actions.last.user
+    assert_equal "Shiny shiny description", @artefact.description
+    assert_equal 2, @artefact.actions.size
+    assert_equal ["create", "update"], @artefact.actions.map(&:action_type)
+    assert_equal user, @artefact.actions.last.user
     assert_equal(
         merge_attributes(DEFAULTS, base_fields, updates),
-        a.actions.last.snapshot
+        @artefact.actions.last.snapshot
     )
   end
 
+  test "saving a task should record the task action" do
+    @artefact.description = "Updated automatically"
+    @artefact.save_as_task!('TaggingUpdater')
+    @artefact.reload
+
+    assert_equal 2, @artefact.actions.size
+    assert_equal ["create", "update"], @artefact.actions.map(&:action_type)
+
+    assert_equal 'TaggingUpdater', @artefact.actions.last.task_performed_by
+    assert_equal nil, @artefact.actions.last.user
+  end
+
   test "saving as a user should record a user action" do
-    a = Artefact.create!(base_fields)
     user = FactoryGirl.create :user
     updates = {description: "Shiny shiny description"}
-    a.description = updates[:description]
-    a.save_as user
-    a.reload
+    @artefact.description = updates[:description]
+    @artefact.save_as user
+    @artefact.reload
 
-    assert_equal "Shiny shiny description", a.description
-    assert_equal 2, a.actions.size
-    assert_equal ["create", "update"], a.actions.map(&:action_type)
-    assert_equal user, a.actions.last.user
+    assert_equal "Shiny shiny description", @artefact.description
+    assert_equal 2, @artefact.actions.size
+    assert_equal ["create", "update"], @artefact.actions.map(&:action_type)
+    assert_equal user, @artefact.actions.last.user
     assert_equal(
         merge_attributes(DEFAULTS, base_fields, updates),
-        a.actions.last.snapshot
+        @artefact.actions.last.snapshot
     )
   end
 
   test "saving as a user with an action type" do
-    a = Artefact.create!(base_fields)
     user = FactoryGirl.create :user
     updates = {description: "Shiny shiny description"}
-    a.description = updates[:description]
-    a.save_as user, action_type: "awesome"
-    a.reload
+    @artefact.description = updates[:description]
+    @artefact.save_as user, action_type: "awesome"
+    @artefact.reload
 
-    assert_equal user, a.actions.last.user
-    assert_equal "awesome", a.actions.last.action_type
+    assert_equal user, @artefact.actions.last.user
+    assert_equal "awesome", @artefact.actions.last.action_type
   end
 
 end