+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## 33.0.0
+
+- Changes `save_as_task` to `save_as_task!` [#364](https://github.com/alphagov/govuk_content_models/pull/364)
+
+## 32.3.1
+
+- Update the Artefact FactoryGirl definition
+
+## 32.3.0
+
+- Add `save_as_task` method to Artefact model [#360](https://github.com/alphagov/govuk_content_models/pull/360)
+
 ## 32.2.0
 
 - Add `homepage_url` attribute to `LocalAuthority` model